requirements.txt
redis>=4.5.0
requests>=2.28.0
flask>=2.3.0
flask-cors>=4.0.0
aiohttp>=3.8.0
loguru>=0.7.0
python-dotenv>=1.0.0
schedule>=1.2.0
gunicorn>=20.1.0  # 新增：生产环境WSGI服务器